Eliminate warning: format specifies type long
Mac OS X defines tv_usec as 'int' instead of 'long':

cckddasd.c:5718:32: warning: format specifies type 'long' but the argument has
      type '__darwin_suseconds_t' (aka 'int') [-Wformat]
                      tbuf+11, tv.tv_usec, dev ? dev->devnum : 0);
                               ^~~~~~~~~~
shared.c:2664:24: warning: format specifies type 'long' but the argument has
      type '__darwin_suseconds_t' (aka 'int') [-Wformat]
            tv.tv_sec, tv.tv_usec, dev ? dev->devnum : 0);
                       ^~~~~~~~~~

+++ cckddasd.c
@@ -5715,7 +5715,7 @@ int             l;
         if (p)
         {
             l = sprintf ((char *)p, "%s" "." "%6.6ld %4.4X:",
-                      tbuf+11, tv.tv_usec, dev ? dev->devnum : 0);
+                      tbuf+11, (long)(tv.tv_usec), dev ? dev->devnum : 0);
             vsprintf ((char *)p + l, msg, vl);
         }
     }
